CK 161009  - Added missing sorted table comments  - Added comments to graph objects  - Fixed Neural Net serialisation and Net - Graph synchronisation  - Changed Network class implementation and refactored ops class  - Added short hand implementation of Neural Net to test artifacts

+++ Src/numl/Serialization/Supervised/NeuralNetwork/EdgeSerializer.cs
@@ -24,8 +24,8 @@ namespace numl.Serialization.Supervised.NeuralNetwork
             {
                 var edge = (Edge)this.Create();
 
-                edge.ParentId = (int)reader.ReadProperty().Value;
-                edge.ChildId = (int)reader.ReadProperty().Value;
+                edge.ParentId = (int)(double)reader.ReadProperty().Value;
+                edge.ChildId = (int)(double)reader.ReadProperty().Value;
                 edge.Weight = (double)reader.ReadProperty().Value;
                 
                 return edge;
